+++ easyfinemap/loci.py
@@ -2,24 +2,27 @@
 
 support three approaches:
 1. identify the independent lead snps by distance only,
-TODO:2. identify the independent lead snps by LD clumping,
-TODO:3. identify the independent lead snps by conditional analysis.
+2. identify the independent lead snps by LD clumping,
+3. identify the independent lead snps by conditional analysis.
 
 than expand the independent lead snps to independent loci by given range.
 merge the overlapped independent loci (optional).
 """
 
 import logging
-import tempfile
+from concurrent.futures import ProcessPoolExecutor
 from pathlib import Path
 from subprocess import PIPE, run
-from typing import List, Optional, Union
+from typing import List, Optional, Tuple, Union
 
 import pandas as pd
+from rich.progress import Progress
 
 from easyfinemap.constant import ColName
+from easyfinemap.ldref import LDRef
 from easyfinemap.tools import Tools
-from easyfinemap.utils import io_in_tempdir, make_SNPID_unique
+from easyfinemap.utils import get_significant_snps, io_in_tempdir, make_SNPID_unique
+from easyfinemap.sumstat import SumStat
 
 
 class Loci:
@@ -29,29 +32,114 @@ class Loci:
         """Initialize the Loci class."""
         self.logger = logging.getLogger("Loci")
         self.plink = Tools().plink
+        self.gcta = Tools().gcta
         self.tmp_root = Path.cwd() / "tmp" / "loci"
         if not self.tmp_root.exists():
             self.tmp_root.mkdir(parents=True)
 
     def identify_indep_loci(
         self,
-        sig_df: pd.DataFrame,
+        sumstats: pd.DataFrame,
+        sig_threshold: float = 5e-8,
+        loci_extend: int = 500,
+        if_merge: bool = False,
+        leadsnp_file: Optional[Path] = None,
+        loci_file: Optional[Path] = None,
+        ldref: Optional[str] = None,
         method: str = "distance",
-        distance: int = 1000000,
-        range: int = 1000000,
-        merge: bool = True,
-    ) -> pd.DataFrame:
+        distance: int = 500,
+        clump_kb: int = 500,
+        clump_r2: float = 0.1,
+        sample_size: Optional[int] = None,
+        cojo_window_kb: int = 10000,
+        cojo_collinear: float = 0.9,
+        diff_freq: float = 0.2,
+        use_ref_EAF: bool = False,
+        threads: int = 1,
+
+    ) -> Tuple[pd.DataFrame, pd.DataFrame]:
         """
         Identify the independent loci.
 
         Parameters
         ----------
-        sig_df : pd.DataFrame
-            The significant snps.
+        sumstats : pd.DataFrame
+            The input summary statistics.
+        sig_threshold : float, optional
+            The pvalue threshold, by default 5e-8
+        loci_extend : int, optional
+            The range to extend the independent lead snps to independent loci, by default 500, unit: kb
+        if_merge : bool, optional
+            Whether to merge the overlapped independent loci, by default False
+        leadsnp_file : Optional[str], optional
+            The output file of the independent lead snps, by default None
+        loci_file : Optional[str], optional
+            The output file of the independent loci, by default None
+        ldref : Optional[str], optional
+            The LD reference file, by default None
         method : str, optional
-            The method to identify the independent loci, by default "distance"
+            The method to identify the independent loci, by default "distance",
+            choose from ["distance", "clumping", "conditional"]
+        distance : int, optional
+            The distance threshold to identify the independent loci, by default 500, unit: kb
+        clump_kb : int, optional
+            The distance threshold for LD clumping, by default 500, unit: kb
+        clump_r2 : float, optional
+            The r2 threshold for LD clumping, by default 0.1
+        sample_size : Optional[int], optional
+            The sample size for conditional analysis, by default None
+        cojo_window_kb : int, optional
+            The distance threshold for conditional analysis, by default 10000
+        cojo_collinear : float, optional
+            The collinear threshold for conditional analysis, by default 0.9
+        diff_freq : float, optional
+            The difference frequency threshold for conditional analysis, by default 0.2
+        use_ref_EAF : bool, optional
+            Whether to use the reference EAF for conditional analysis, by default False
+        threads : int, optional
+            The number of threads, by default 1
+
+        Returns
+        -------
+        Tuple[pd.DataFrame, pd.DataFrame]
+            The independent lead snps and independent loci.
         """
-        raise NotImplementedError
+        sumstats = SumStat(sumstats)
+        sumstats = sumstats.standarize()
+        if method == "distance":
+            sig_df = get_significant_snps(sumstats, sig_threshold)
+            lead_snp = self.indep_snps_by_distance(sig_df, distance)
+        elif method == "clumping":
+            clump_p1 = sig_threshold
+            if ldref is not None:
+                lead_snp = self.indep_snps_by_ldclumping(sumstats, ldref, clump_p1, clump_kb, clump_r2)
+            else:
+                raise ValueError(f"Please provide the ldref file for method: {method}")
+        elif method == "conditional":
+            if ldref is None:
+                raise ValueError("Please provide the ldref file for conditional analysis.")
+            if sample_size is None:
+                raise ValueError("Please provide the sample size for conditional analysis.")
+            else:
+                lead_snp = self.indep_snps_by_conditional(
+                    sumstats,
+                    ldref,
+                    sample_size,
+                    sig_threshold,
+                    cojo_window_kb,
+                    cojo_collinear,
+                    diff_freq,
+                    use_ref_EAF,
+                    threads,
+                )
+        else:
+            raise ValueError(f"Unsupported method: {method}")
+        loci = self.leadsnp2loci(lead_snp, loci_extend, if_merge)
+        if loci_file:
+            loci.to_csv(loci_file, sep="\t", index=False)
+        if leadsnp_file:
+            lead_snp.to_csv(leadsnp_file, sep="\t", index=False)
+        return lead_snp, loci
 
     @staticmethod
     def merge_overlapped_loci(loci_df: pd.DataFrame):
@@ -88,7 +176,7 @@ class Loci:
         return result
 
     @staticmethod
-    def indep_snps_by_distance(sig_df: pd.DataFrame, distance: int = 500000) -> pd.DataFrame:
+    def indep_snps_by_distance(sig_df: pd.DataFrame, distance: int = 500) -> pd.DataFrame:
         """
         Identify the independent snps by distance only.
 
@@ -97,7 +185,7 @@ class Loci:
         sig_df : pd.DataFrame
             The significant snps.
         distance : int, optional
-            The distance threshold, by default 1000000
+            The distance threshold, by default 500, unit: kb
 
         Returns
         -------
@@ -106,6 +194,7 @@ class Loci:
         """
         sig_df.sort_values(ColName.P, inplace=True)
         lead_snp = []
+        distance = distance * 1000
         while len(sig_df):
             lead_snp.append(sig_df.iloc[[0]])
             sig_df = sig_df[
@@ -120,7 +209,7 @@ class Loci:
 
     @staticmethod
     def indep_snps_by_ldclumping(
-        sig_df: pd.DataFrame, ldref: str, clump_p1: float = 5e-8, clump_kb: int = 500000, clump_r2: float = 0.1
+        sig_df: pd.DataFrame, ldref: str, clump_p1: float = 5e-8, clump_kb: int = 500, clump_r2: float = 0.1
     ) -> pd.DataFrame:
         """
         Identify the independent snps by LD clumping.
@@ -134,7 +223,7 @@ class Loci:
         clump_p1 : float, optional
             The p1 threshold, by default 5e-8
         clump_kb : int, optional
-            The kb threshold, by default 500000
+            The kb threshold, by default 500, unit: kb
         clump_r2 : float, optional
             The r2 threshold, by default 0.1
 
@@ -216,20 +305,106 @@ class Loci:
             return clump_snps
 
     @staticmethod
-    def indep_snps_by_conditional(sig_df: pd.DataFrame, ld_df: pd.DataFrame, r2_threshold: float = 0.8) -> pd.DataFrame:
-        """Identify the independent snps by conditional analysis."""
-        raise NotImplementedError
-
-    def cojo_per_chr(self, sig_df: pd.DataFrame, ldref: str, temp_dir: Optional[str] = None) -> pd.DataFrame:
+    def indep_snps_by_conditional(
+        sumstats: pd.DataFrame,
+        ldref: str,
+        sample_size: int,
+        sig_threshold: float = 5e-8,
+        cojo_window_kb: int = 10000,
+        cojo_collinear: float = 0.9,
+        diff_freq: float = 0.2,
+        use_ref_EAF: bool = False,
+        threads: int = 1,
+    ) -> pd.DataFrame:
         """
-        Conditional analysis per chromosome.
+        Identify the independent snps by conditional analysis.
 
         Parameters
         ----------
-        sig_df : pd.DataFrame
-            The significant snps.
+        sumstats : pd.DataFrame
+            The summary statistics.
         ldref : str
             The LD reference file, (plink bfile format, containing wildcard {chrom}), e.g. EUR.chr{chrom}.
+        sample_size : int
+            The sample size.
+        sig_threshold : float, optional
+            The significance threshold, by default 5e-8
+        cojo_window : int, optional
+            The cojo window, by default 10000, in kb
+        cojo_collinear : float, optional
+            The cojo collinear, by default 0.9
+        diff_freq : float, optional
+            The difference frequency, by default 0.2
+        use_ref_EAF : bool, optional
+            Whether to use the reference EAF, by default False
+        threads : int, optional
+            The number of threads, by default 1
+        TODO: accelerate the process by running cojo in loci identified by distance
+        """
+        logger = logging.getLogger('COJO')
+        if not use_ref_EAF and ColName.EAF not in sumstats.columns:
+            raise ValueError(f"{ColName.EAF} is not in the sumstats, please set use_ref_EAF to True")
+        sig_df = sumstats[sumstats[ColName.P] <= sig_threshold]
+        logger.debug(f"Number of significant snps: {len(sig_df)}")
+        logger.debug(f"Number of chromosomes: {len(sig_df[ColName.CHR].unique())}")
+        args_list = []
+        for chrom in sig_df[ColName.CHR].unique():
+            in_df = sumstats[sumstats[ColName.CHR] == chrom]
+            args_list.append(
+                (
+                    in_df,
+                    ldref.format(chrom=chrom),
+                    sample_size,
+                    cojo_window_kb,
+                    cojo_collinear,
+                    diff_freq,
+                    sig_threshold,
+                    use_ref_EAF,
+                )
+            )
+        with ProcessPoolExecutor(max_workers=threads) as executor:
+            loci = Loci()
+            cojo_snps = executor.map(loci.cojo_slct, *zip(*args_list))
+            results = []
+            with Progress(auto_refresh=False) as progress:
+                task = progress.add_task("Run cojo-slct", total=len(args_list))
+                for _ in executor.map(loci.cojo_slct, *zip(*args_list)):
+                    progress.update(task, advance=1)
+                    progress.refresh()
+                    results.append(_)
+        cojo_snps = pd.concat(cojo_snps, axis=0, ignore_index=True)
+        return cojo_snps
+
+    @io_in_tempdir(dir="./tmp/loci")
+    def cojo_slct(
+        self,
+        sumstats: pd.DataFrame,
+        ldref: str,
+        sample_size: int,
+        cojo_window_kb: int = 10000,
+        cojo_collinear: float = 0.9,
+        diff_freq: float = 0.2,
+        sig_threshold: float = 5e-8,
+        use_ref_EAF: bool = False,
+        temp_dir: Optional[str] = None,
+    ) -> pd.DataFrame:
+        """
+        Conditional analysis for input sumstatistics.
+
+        Parameters
+        ----------
+        sumstats : pd.DataFrame
+            The input sumstatistics, from same chromosome or locus.
+        ldref : str
+            The LD reference file, (plink bfile format, containing wildcard {chrom}), e.g. EUR.chr{chrom}.
+        sample_size : int
+            The sample size of the input sumstatistics.
+        cojo_window : int, optional
+            The cojo window, by default 10000, unit: kb
+        cojo_collinear : float, optional
+            The cojo collinear, by default 0.9
+        diff_freq : float, optional
+            The difference frequency, by default 0.2
         temp_dir : Optional[str], optional
             The temporary directory, by default None
 
@@ -238,53 +413,91 @@ class Loci:
         pd.DataFrame
             The conditional snps.
         """
-        chrom = sig_df[ColName.CHR].unique()[0]
-        cojo_p_file = f"{temp_dir}/cojo_p_{chrom}.txt"
-        sig_df[[ColName.SNPID, ColName.P]].to_csv(cojo_p_file, sep="\t", index=False)
-        cojo_outfile = f"{temp_dir}/cojo_{chrom}.cojo"
+        chrom = sumstats[ColName.CHR].unique()[0]
+        cojo_input = sumstats.copy()
+        ld = LDRef()
+        cojo_input = ld.intersect(sumstats, ldref, f"{temp_dir}/cojo_input_{chrom}", use_ref_EAF)
+        cojo_input[ColName.N] = sample_size
+        cojo_input = cojo_input[
+            [ColName.SNPID, ColName.EA, ColName.NEA, ColName.EAF, ColName.BETA, ColName.SE, ColName.P, ColName.N]
+        ]
+        cojo_input.rename(
+            columns={
+                ColName.SNPID: "SNP",
+                ColName.EA: "A1",
+                ColName.NEA: "A2",
+                ColName.EAF: "freq",
+                ColName.BETA: "b",
+                ColName.SE: "se",
+                ColName.P: "p",
+                ColName.N: "N",
+            },
+            inplace=True,
+        )
+        cojo_p_file = f"{temp_dir}/cojo_input_{chrom}.ma"
+        cojo_input.to_csv(cojo_p_file, sep=" ", index=False)
+        cojo_outfile = f"{temp_dir}/cojo_{chrom}.slct"
         cmd = [
-            self.plink,
+            self.gcta,
             "--bfile",
-            ldref.format(chrom=chrom),
+            ldref,
             "--cojo-file",
             cojo_p_file,
             "--cojo-slct",
             "--cojo-p",
-            "1",
+            str(sig_threshold),
+            "--cojo-wind",
+            str(cojo_window_kb),
+            "--cojo-collinear",
+            str(cojo_collinear),
+            "--diff-freq",
+            str(diff_freq),
             "--out",
-            f"{temp_dir}/cojo_{chrom}",
+            cojo_outfile,
         ]
         res = run(cmd, stdout=PIPE, stderr=PIPE, universal_newlines=True)
         if res.returncode != 0:
             self.logger.error(res.stderr)
             raise RuntimeError(res.stderr)
         else:
-            cojo_snps = pd.read_csv(cojo_outfile, delim_whitespace=True, usecols=["SNP"])
-            cojo_snps = cojo_snps["SNP"].to_list()
-            cojo_snps = sig_df[sig_df[ColName.SNPID].isin(cojo_snps)]
+            cojo_snps = pd.read_csv(
+                f"{cojo_outfile}.jma.cojo", delim_whitespace=True, usecols=["SNP", "bJ", "bJ_se", "pJ"]
+            )
+            cojo_snps.rename(
+                columns={
+                    "SNP": ColName.SNPID,
+                    "bJ": ColName.COJO_BETA,
+                    "bJ_se": ColName.COJO_SE,
+                    "pJ": ColName.COJO_P,
+                },
+                inplace=True,
+            )
+            cojo_snps = cojo_snps[cojo_snps[ColName.COJO_P] <= sig_threshold]
+            cojo_snps = sumstats.merge(cojo_snps, on=ColName.SNPID, how="inner")
             return cojo_snps
 
     @staticmethod
-    def leadsnp2loci(sig_df: pd.DataFrame, range: int = 500000, if_merge: bool = True) -> pd.DataFrame:
+    def leadsnp2loci(lead_snps: pd.DataFrame, range: int = 500, if_merge: bool = False) -> pd.DataFrame:
         """
         Expand the independent lead snps to independent loci by given range.
 
         Parameters
         ----------
-        sig_df : pd.DataFrame
+        lead_snps : pd.DataFrame
             The independent lead snps.
         range : int, optional
-            The range, by default 500000
+            The range, by default 500, unit: kb
         if_merge : bool, optional
-            Whether merge the overlapped loci, by default True
+            Whether merge the overlapped loci, by default False
 
         Returns
         -------
         pd.DataFrame
             The independent loci.
         """
-        loci_df = sig_df.copy()
-        loci_df = make_SNPID_unique(loci_df)
+        loci_df = lead_snps.copy()
+        # loci_df = make_SNPID_unique(loci_df)
+        range = range * 1000
         loci_df = loci_df[[ColName.CHR, ColName.BP, ColName.P, ColName.SNPID]]
         loci_df.columns = [ColName.CHR, ColName.LEAD_SNP_BP, ColName.LEAD_SNP_P, ColName.LEAD_SNP]  # type: ignore
         loci_df[ColName.START] = loci_df[ColName.LEAD_SNP_BP] - range
@@ -292,5 +505,7 @@ class Loci:
         loci_df[ColName.END] = loci_df[ColName.LEAD_SNP_BP] + range
         loci_df = loci_df[ColName.loci_cols].copy()
         if if_merge:
+            if ColName.COJO_BETA in loci_df.columns:
+                logging.warning("The loci identified by cojo may not need merge.")
             loci_df = Loci.merge_overlapped_loci(loci_df)
         return loci_df